+++ queries/_select-puzzle-resources-preview_full-by-puzzle_id.sql
@@ -4,4 +4,6 @@ join Puzzle as p1 on (p1.id = pi.original)
 JOIN PuzzleFile AS pf ON (pf.puzzle = p1.id) -- Get the original
 WHERE p.puzzle_id = :puzzle_id
 AND pf.name = 'preview_full'
+AND p.status not in (0, -1, -2, -10, -11, -12, -20, -21)
+-- NEEDS_MODERATION FAILED_LICENSE NO_ATTRIBUTION DELETED_LICENSE DELETED_INAPT DELETED_OLD SUGGESTED SUGGESTED_DONE
 ;